Finished project

+++ Domain/Domains/Student.cs
@@ -16,7 +16,7 @@ public sealed class Student : Entity
     public string Email { get; set; }
 
     [Required, Range(1, 5)]
-    public int AcademicPerformance { get; set; }
+    public double AcademicPerformance { get; set; }
 
     [Required]
     public int GroupId { get; set; }
